+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Only displaying _Ethernet_ kind interfaces
 - Displaying the type of ethernet adapter on macOS
 - Always reselecting the last selected Descriptor when reinspecting an Entity
+- [Moved ProtocolInterface selection to the Settings](https://github.com/christophe-calmejane/Hive/issues/58)
 
 ### Fixed
 - Possible deadlock when trying to match stream formats